Add ZeroMQ functionality to send logs in JSON format
Added support to send logs in JSON format via ZeroMQ.  Added Bro
scripting support to enable a user to configure the destination
hostname and port number, and the Bro log IDs to send.  Added
a "--with-zmq" option to the "configure" script to help build
the plugin when zmq cannot be found automatically.

Also updated the test baseline.